+++ samples/java-mule/src/main/app/mule-swagger-integration.xml
@@ -30,8 +30,8 @@ http://www.mulesoft.org/schema/mule/json http://www.mulesoft.org/schema/mule/jso
 	<spring:beans>
 		<spring:bean id="bootstrap" init-method="start" lazy-init="false"
 			class="com.wordnik.swagger.sample.Bootstrap" />
-		<spring:bean id="apiListingJSON"
-			class="com.wordnik.swagger.mule.ApiListingJSON" />
+		<spring:bean id="muleApiListing"
+			class="com.wordnik.swagger.mule.MuleApiListing" />
 		<spring:bean id="swaggerConfig"
 			class="com.wordnik.swagger.jaxrs.config.BeanConfig">
 			<spring:property name="resourcePackage" value="com.wordnik.swagger.sample.resource" />
@@ -73,7 +73,7 @@ http://www.mulesoft.org/schema/mule/json http://www.mulesoft.org/schema/mule/jso
               <spring-object bean="userResource" />
             </component>
             <component>
-							<spring-object bean="apiListingJSON" />
+							<spring-object bean="muleApiListing" />
 						</component>
 					</jersey:resources>
